Rename MetaDataLoaderConnectionAdapter's package
