+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/metadata/schema/loader/physical/adapter/handler/dialect/MariaDBDatabaseMetaDataDialectHandler.java
@@ -15,21 +15,21 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.metadata.schema.loader.physical.jdbc.handler.dialect;
+package org.apache.shardingsphere.infra.metadata.schema.loader.physical.adapter.handler.dialect;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.apache.shardingsphere.infra.metadata.schema.loader.physical.jdbc.handler.DatabaseMetaDataDialectHandler;
+import org.apache.shardingsphere.infra.metadata.schema.loader.physical.adapter.handler.DatabaseMetaDataDialectHandler;
 import org.apache.shardingsphere.sql.parser.sql.common.constant.QuoteCharacter;
 
 import java.util.Properties;
 
 /**
- * Database meta data dialect handler of MySQL.
+ * Database meta data dialect handler of MariaDB.
  */
 @Getter
 @Setter
-public final class MySQLDatabaseMetaDataDialectHandler implements DatabaseMetaDataDialectHandler {
+public final class MariaDBDatabaseMetaDataDialectHandler implements DatabaseMetaDataDialectHandler {
     
     private Properties props;
     
@@ -40,6 +40,6 @@ public final class MySQLDatabaseMetaDataDialectHandler implements DatabaseMetaDa
     
     @Override
     public String getType() {
-        return "MySQL";
+        return "MariaDB";
     }
 }
