+++ shardingsphere-infra/shardingsphere-infra-common/src/test/java/org/apache/shardingsphere/infra/metadata/schema/loader/physical/adapter/handler/dialect/PostgreSQLDatabaseMetaDataDialectHandlerTest.java
@@ -15,10 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.metadata.schema.loader.physical.jdbc.handler.dialect;
+package org.apache.shardingsphere.infra.metadata.schema.loader.physical.adapter.handler.dialect;
 
-import org.apache.shardingsphere.infra.database.type.dialect.SQL92DatabaseType;
-import org.apache.shardingsphere.infra.metadata.schema.loader.physical.jdbc.handler.AbstractDatabaseMetaDataDialectHandlerTest;
+import org.apache.shardingsphere.infra.database.type.dialect.PostgreSQLDatabaseType;
+import org.apache.shardingsphere.infra.metadata.schema.loader.physical.adapter.handler.AbstractDatabaseMetaDataDialectHandlerTest;
 import org.apache.shardingsphere.sql.parser.sql.common.constant.QuoteCharacter;
 import org.junit.Test;
 
@@ -28,25 +28,25 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.when;
 
-public final class SQL92DatabaseMetaDataDialectHandlerTest extends AbstractDatabaseMetaDataDialectHandlerTest {
+public final class PostgreSQLDatabaseMetaDataDialectHandlerTest extends AbstractDatabaseMetaDataDialectHandlerTest {
     
     @Test
     public void assertGetSchema() throws SQLException {
         when(getConnection().getSchema()).thenReturn(DATABASE_NAME);
-        String sql92Schema = getSchema(new SQL92DatabaseType());
-        assertThat(sql92Schema, is(DATABASE_NAME));
+        String postgresqlSchema = getSchema(new PostgreSQLDatabaseType());
+        assertThat(postgresqlSchema, is(DATABASE_NAME));
     }
     
     @Test
     public void assertFormatTableNamePattern() {
-        String sql92TableNamePattern = formatTableNamePattern(new SQL92DatabaseType());
-        assertThat(sql92TableNamePattern, is(TABLE_NAME_PATTERN));
+        String postgresqlTableNamePattern = formatTableNamePattern(new PostgreSQLDatabaseType());
+        assertThat(postgresqlTableNamePattern, is(TABLE_NAME_PATTERN));
     }
     
     @Test
     public void assertGetQuoteCharacter() {
-        QuoteCharacter sql92QuoteCharacter = getQuoteCharacter(new SQL92DatabaseType());
-        assertThat(sql92QuoteCharacter.getStartDelimiter(), is("\""));
-        assertThat(sql92QuoteCharacter.getEndDelimiter(), is("\""));
+        QuoteCharacter postgresqlQuoteCharacter = getQuoteCharacter(new PostgreSQLDatabaseType());
+        assertThat(postgresqlQuoteCharacter.getStartDelimiter(), is("\""));
+        assertThat(postgresqlQuoteCharacter.getEndDelimiter(), is("\""));
     }
 }
